backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
pydantic
openai
python-dotenv
httpx
pytest
pytest-asyncio
pytest-cov
websockets 